Support for anchore sbom and scan action with evidence integration
Add example workflows showcasing the support of anchore scan and sbom
actions, and attaching the reports generated as evidence to the
associated artifacts in Jfrog Artifactory. This commit also has helper
scripts to convert the sarif and json files generated by anchore tools
to readme which can be optionally attached with evidence.